refactor messagepool provider out of the main messagepool implementation
